# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: AI App Template API
  version: 0.3.0
  description: Server Actions (actions/*.ts) are typed RPC — their contract is TypeScript + the shared Zod schemas in lib/validations, enforced at compile time; this OpenAPI documents the HTTP route-handler surface only. Auth.js routes under /api/auth/[...nextauth] are framework-owned and omitted. This file is GENERATED from Zod (the single source of truth) via `pnpm openapi:generate` — never hand-edit it; a smoke drift gate fails if it is stale.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: your-domain.com
...
components:
  securitySchemes:
    cronBearer:
      type: http
      scheme: bearer
      description: "CRON_SECRET bearer token. Fails closed (401) when the secret is unset. Sent as `Authorization: Bearer <CRON_SECRET>`."
    apiKey:
      type: http
      scheme: bearer
      description: "E267 API key. Sent as `Authorization: Bearer sk_<prefix>_<secret>`. Resolves to the owning user; the key's scopes gate access (read → GET, write → POST). 401 if missing/invalid/revoked, 403 if the scope is absent."
  schemas:
...
    CreateItemInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
      required:
        - title
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
        - updatedAt
    ItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
      required:
        - items
    ItemCreateResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/Item"
      required:
        - item
    ItemsCsvExport:
      type: string
      description: "RFC-4180 CSV with columns: id, title, createdAt, updatedAt."
...
paths:
...
  /api/v1/items:
    get:
      summary: List the API key owner's items
      description: Lists the items owned by the user the API key resolves to (newest first). Requires the `read` scope. Bumps the key's lastUsedAt on success.
      tags:
        - items
      security:
        - apiKey: []
      responses:
        "200":
          description: The key owner's items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "401":
          description: Missing, malformed, invalid, or revoked API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The API key lacks the required `read` scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create an item
      description: Creates an item owned by the user the API key resolves to. Requires the `write` scope. The body is validated by the shared CreateItemInput Zod schema. Bumps the key's lastUsedAt on success.
      tags:
        - items
      security:
        - apiKey: []
      requestBody:
        description: The item to create.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemInput"
      responses:
        "201":
          description: The created item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCreateResponse"
        "400":
          description: Invalid JSON body or schema validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, malformed, invalid, or revoked API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The API key lacks the required `write` scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/export/items:
    get:
      summary: Export the API key owner's items as CSV
      description: "Downloads all items owned by the user the API key resolves to as an RFC-4180 CSV file. Requires the `read` scope. Returns `Content-Disposition: attachment` so browsers trigger a file download. Bumps the key's lastUsedAt on success."
      tags:
        - items
      security:
        - apiKey: []
      responses:
        "200":
          description: CSV attachment of the key owner's items.
          content:
            text/csv:
              schema:
                $ref: "#/components/schemas/ItemsCsvExport"
        "401":
          description: Missing, malformed, invalid, or revoked API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The API key lacks the required `read` scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...